Catch unicode conversion errors
Many unicode characters are more than 1 byte long (e.g. emoji). If one
tries to take a slice on a string containing such characters that
would otherwise 'break' the character, a panic is thrown.

The `from_hex` function had a bug that checked if the first 2 cahrs of a
string are '0x' when converting from string to a byte string. Passing an
emoji to this function would cause a panic.

Since any non-ascii characters aren't valid hexadecimal anyway, we can
check that all characters are actually ascii before doing any slice
checks.

A test was added that catches this case and confirms that the bug is
fixed.