feat: ledger compatability (#54)
This PR is to add no_std support. This is in-line to ensure that
tari_utilities can compile for nonstandard tier 1 rust targets. This is
required for the push to ensure we can create a ledger-compatible
library and allow us to run some of the handy utilities from this crate.

The changes can be summarized in the following:

- Made the library `no_std`
- Upgraded borsch to version 0.10, up from 0.9
- Added feature flags to hide crates and features that are not possible
on `no_std` targets

Feature flags added:

- `std` - This will allow epoch_time, encoding and safe_array
- `zero` - This will include zeroize support
- `serialize` - This will include support for serde
- `borsh` - This will include support for borsh
- `default` - This will include all feature flags and allow all
utilities.

---------

Co-authored-by: stringhandler <stringhandler@gmail.com>