chore: update CI and features (#71)
This PR updates CI and relaxes features.

When running CI, a mix of nightly and stable toolchains is used. This PR
moves to the stable toolchain for all jobs except formatting, which
requires nightly. It also updates the CI actions to more updated and
supported versions.

It also relaxes feature gating and simplifies dependencies while
retaining `no_std` compatibility.